changed 'infrastructure' to 'software', line 104
The word 'infrastructure' was used to describe a web browser. This was changed to 'software' as it is a more accurate description. 

Infrastructure refers to hardware, a web browser is a software application. The sentence now  reads, 'Don't confuse the software (e.g., the browser) with the service (e.g., the search engine).'